brush: refactor Map splitting it into OAMLayer and TilemapLayer
