+++ brush/src/components/map/index.js
@@ -1,122 +1,36 @@
 import React, { useContext, Fragment, useState } from 'react'
 import PropTypes from 'prop-types'
 import { Stage, Layer } from 'react-konva'
-import {
-  defaultTo,
-  find,
-  pipe,
-  prop,
-  range,
-} from 'ramda'
 import MapFooter from '../MapFooter'
-import PointTile from './point/PointTile'
-import PointOam from './point/PointOam'
 import HighlightCoordinates from './HighlightCoordinates'
 import VisionContext from '../../context/VisionContext'
-
-const listCoordinates = (height, width) => {
-  const rangeHeight = range(0, height)
-  const rangeWidth = range(0, width)
-
-  const coordinates = rangeWidth.reduce((acc, i) => {
-    rangeHeight.forEach(j => acc.push([i, j]))
-    return acc
-  }, [])
-
-  return coordinates
-}
-
-const getTileNameFromScheme = tileValue => pipe(
-  find(({ ids }) => ids.includes(tileValue)),
-  defaultTo({ name: 'unknown' }),
-  prop('name')
-)
+import TilemapLayer from './TilemapLayer'
+import OAMLayer from './OAMLayer'
 
 const Map = ({ highlightCoordinates }) => {
+  const { vision } = useContext(VisionContext) // workaround because a limitation in react-konva (https://github.com/konvajs/react-konva/issues/349)
   const {
     infos: {
       tilemap: {
         height,
-        scheme,
         width,
       },
     },
-    oam,
-    tilemap,
-  } = useContext(VisionContext).vision
+  } = vision
 
   const [selectedPointInfos, setSelectedPointInfos] = useState(null)
-
-  const getPoint = (x, y) => {
-    const tileValue = tilemap[x + (y * width)]
-
-    let tileName
-    if (tileValue === 0x00) {
-      tileName = 'empty'
-    } else {
-      tileName = getTileNameFromScheme(tileValue)(scheme)
-    }
-
-    return (<PointTile
-      key={`${x} ${y}`}
-      tileName={tileName}
-      tileValue={tileValue}
-      showPointInfosHandle={setSelectedPointInfos}
-      x={x}
-      y={y}
-    />)
-  }
-
-  const oamList = oam
-    .map(oamEntry => oamEntry.data)
-    .map(({
-      kind,
-      xStage1,
-      xStage2,
-      xStage3,
-      yStage1,
-      yStage2,
-      yStage3,
-    }) =>
-      [
-        <PointOam
-          key={`${xStage1} ${yStage1} 1`}
-          oamId={kind}
-          stage={1}
-          showPointInfosHandle={setSelectedPointInfos}
-          x={xStage1}
-          y={yStage1}
-        />,
-        <PointOam
-          key={`${xStage2} ${yStage2} 2`}
-          oamId={kind}
-          stage={2}
-          showPointInfosHandle={setSelectedPointInfos}
-          x={xStage2}
-          y={yStage2}
-        />,
-        <PointOam
-          key={`${xStage3} ${yStage3} 3`}
-          oamId={kind}
-          stage={3}
-          showPointInfosHandle={setSelectedPointInfos}
-          x={xStage3}
-          y={yStage3}
-        />,
-      ])
-
-  const tiles = listCoordinates(height, width)
-    .map(([x, y]) => getPoint(x, y))
 
   return (
     <Fragment>
       <Stage width={width * 4} height={height * 4}>
-        <Layer>
-          {tiles}
-        </Layer>
-        <Layer>
-          {oamList}
-        </Layer>
+        <TilemapLayer
+          setSelectedPointInfos={setSelectedPointInfos}
+          vision={vision}
+        />
+        <OAMLayer
+          setSelectedPointInfos={setSelectedPointInfos}
+          vision={vision}
+        />
         <Layer>
           <HighlightCoordinates
             coordinates={highlightCoordinates}
